==> Instacart Basket Analysis/Scripts/1 Data importing and descriptive analysis.ipynb <==
# Importing Imporatant packages

#Import Libraries
import pandas as pd
import numpy as np
import os

path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'

df= pd.read_csv(os.path.join(path,'02 data', 'Original Data', '4.3 orders.csv'),index_col=False)

df.head()

##Import the “orders.csv” data set into your notebook using the os library and the vars_list shortcut, omitting the “eval_set” column in your final dataframe.


vars_list= ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

df=pd.read_csv(os.path.join(path,'02 data', 'original data', '4.3 orders.csv'), usecols=vars_list)

## Print the first 5 rows of your dataframe.
df.head()

##Import the “products.csv” file into Jupyter as df_prods
df_prods= pd.read_csv(os.path.join(path,'02 data', 'Original Data', '4.3 products.csv'),index_col=False)

##Print the first 20 rows and the last 35 rows of the df_prods dataframe

df_prods.head(20)

df_prods.tail(35)

##Print the names of the columns in the df_prods dataframe.

df_prods.columns

##Print the number of rows and columns in the df_prods dataframe.
df_prods.shape

df_prods.describe()

df_prods.dtypes
