==> Instacart Basket Analysis/Scripts/2 Data Wrangling & Subsetting.ipynb <==
# Importing Libraries

import pandas as pd
import numpy as np
import os

##Importating orders.csv
path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'
df_ords= pd.read_csv(os.path.join(path,'02 data', 'Original Data', '4.3 orders.csv'),index_col=False)

##Importing products.csv
df_prods= pd.read_csv(os.path.join(path,'02 data', 'Original Data', '4.3 products.csv'),index_col=False)


#Dropping orders.csv eval_set column
df_ords=df_ords.drop(columns = ['eval_set'])

##Finding missing values in the days_since_prior_order column

df_ords['days_since_prior_order'].value_counts(dropna = False)

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

##Changing order_id into a string data type
df_ords['order_id'] = df_ords['order_id'].astype('str')

##Importing df_dep dataframe

df_dep= pd.read_csv(os.path.join(path,'02 data', 'Original Data', '4.4 departments.csv'),index_col=False)


#transposing df_dep 
df_dep.T
df_dep_t=df_dep.T
df_dep_t.reset_index()


##Making df_dep_t_new starting from row 1 so we can delete the header row
new_header=df_dep_t.iloc[0]
df_dep_t_new = df_dep_t[1:]
df_dep_t_new.columns = new_header
df_dep_t_new

##Making df_dep_t_new into a data dictionary

data_dict = df_dep_t_new.to_dict('index')

data_dict

##Subsetting df_snacks by department 19 as a criterion

df_snacks =  df_prods[df_prods['department_id']==19]

##Subsetting using the loc function
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

##Subsetting using the loc and isin function

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

##Part 4.4 Procedure Answers


df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords['order_number'] = df_ords['order_number'].astype('str')
df_ords.dtypes

##Changed user_id into a string since it does not get value as a number. These are unique identifiers that do not provide much value in their number data types


##changed the order_hour_of_day' to order_time_hour_of_day to show it is a time variable
df_ords.rename(columns = {'order_hour_of_day' : 'order_time_hour_of_day'}, inplace = True)

df_ords



##4. Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

df_ords['order_time_hour_of_day'].value_counts(dropna = False)

## 10 am is found to be the most popular time of the day with the most count values

##5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.

print(data_dict.get('4'))

##we will refer to the datadict variable define earlier to understand the number 4 -> which is the produce department

##6. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

df_breakfast= df_prods[df_prods['department_id']==14]
df_breakfast
##subsetting based on department 14 which is the breakfast department

##7. They’d also like to see details about products that customers might use to throw dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

#alcohol = department 5
#deli = department 20
#beverages = department 7
#meat/seafood = department 12


#will the subsetting technique, but also use isin in order to make a list for python to search through
df_dinner_list = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_list

##8. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

df_dinner_list.shape

#7650 rows, 5 columns

##9. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

## wiill use the subsetting technique with a criterion of user_id=1
df_user_id_1= df_ords[df_ords['user_id']=='1']
##since earlier, we changed the user_id data type to a string, a quotation mark around '1' is necessary


df_user_id_1

##10. You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?


df_user_id_1.describe()

##Basic stats include a count of 11 orders, that usually occur on day 2 and usually occurs around 10 am

df_dep_t

#Exporting df_ords
df_ords.to_csv(os.path.join(path,'02 Data','Prepared Data','orders_wrangled.csv'))


#Exporting df_dep_t_new

df_dep_t_new.to_csv(os.path.join(path,'02 Data','Prepared Data','departments_wrangled.csv'))


##For both of these pathways, it was important to send the csv files to the prepared data folder vs. the original data folders
