==> Instacart Basket Analysis/Scripts/3 Data Consistency Checks.ipynb <==
## Importing Libraries

##Importing important packages

import pandas as pd
import numpy as np
import os


##Importing path, and csv files of products

path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'

# To import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', '4.3 products.csv'), index_col=False)

# To import orders_wrangled.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', '4.4 orders_wrangled.csv'), index_col=False)


#Creating a dataframe

df_test= pd.DataFrame()

#Creating a mixed type column

df_test['mix']=['a','b',1,True]

#Checking for mixed-types

for col in df_prods.columns.tolist():
    weird = (df_prods[col].map(type) != type(df_prods[col].iloc[0])).any()
    if weird:
        print(col)

#Function to count any missing values 

df_prods.isnull().sum()


#Subsetting df_prods for only the 16 missing product_name

df_nan =  df_prods[df_prods['product_name'].isnull()==True]
df_nan

#Subsetting df_prods_clean to remove missing values by setting .isnull() to false

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]


#Finding duplicate rows of the df_prods_clean dataframe

df_dups = df_prods_clean[df_prods_clean.duplicated()]


#Making new df using drop_duplicates function
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups

#This new Df_prods_clean_no_dups is exactly 5 less than the df_prods_clean so we know it was done correctly
#Will rename this into "products_checked.csv" when exporting since df_prods_clean_no_dups can be long

## 4.5 Task Section Header

# 2. Run the df.describe() function on your df_ords dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.

df_ords.describe()




orders_day_of_week: The min and max are from 0 to 6, which represent Sunday through Saturday. These should all be whole numbers and is reforced by the 50% mid point being exactly 3. This means that there are no float variables that may skew the data

order_time_hour_of_day: The min and max times are 0 and 23. This is in military time and it make sense that there is not a 24 as that is equivalent to 0 for time. 

days_since_prior_order: All of these values are (+) values, along with the other columns as there cannot be a negative time. 

# 3. Check for mixed-type data in your df_ords dataframe.


# This code was copied from the CareerFoundry website. It continued to receive an error message that the DataFrame.applymap is deprecate and possibly going to be removed upon futuer updates. Online research said to use map instead of applymap to avoid this
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)     



# 4. If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

#There was no mixed types found within the column. However, there are many nulls when using the isnull() function on df_ords

#5. Run a check for missing values in your df_ords dataframe.
#In a markdown cell, report your findings and propose an explanation for any missing values you find.

    
df_ords.isnull().sum()



5. 
isnull() finds 206209 null values in this column. This data is an integer and imputation can be applicable. We will imputate using the mean value.

#6. Address the missing values using an appropriate method

df_ords['days_since_prior_order'] = df_ords['days_since_prior_order'].fillna(df_ords['days_since_prior_order'].mean())
df_ords



6. 
This data in days_since_prior_order is an integer and imputation can be applicable. We will imputate using the mean value since it is provides a middle average of the days data.


# 7.Run a check for duplicate values in your df_ords data.
# In a markdown cell, report your findings and propose an explanation for any duplicate values you find.

# 8. Address the duplicates using an appropriate method.
# In a markdown cell, explain why you used your method of choice.

    
df_ords_dups = df_ords[df_ords.duplicated()]
df_ords_dups.head()

7 and 8. From the duplicated() function, it shows that there are no duplicated rows, so no further action needs to be done. If there were, then we would go through process of deleting them through the drop_duplicates() function

## Exporting new csv files



#Exporting df_ords
df_ords.to_csv(os.path.join(path,'02 Data','Prepared Data','4.5 orders_checked.csv'))


#Exporting df_dep_t_new

df_prods.to_csv(os.path.join(path,'02 Data','Prepared Data','4.5 prods_checked.csv'))
