==> Instacart Basket Analysis/Scripts/4 Combining & Exporting Data Part A.ipynb <==
## Combining & Exporting Data

# Import libraries

import pandas as pd
import numpy as np
import os

##Importing path, and csv files of products

path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'

# To import df_ords_prior

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', '4.6 orders_products_prior.csv'), index_col=False)

## 4. Check the dimensions of the imported dataframe.
##Shape of the data frame
## (32434489 rows, 4 columns)
df_ords_prior.shape

df_ords_prior.head()

## Merging df_ords_prior with df_ords

## Importing df_ords data frame from Exercise 4.5 ("4.5 orders_checked".csv")

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', '4.5 orders_checked.csv'), index_col=False)

df_ords.head()

## Checking shape of the df_ords data frame
## (3421083 rows, 8 columns)
df_ords.shape

## Will use the merge() function to combine the new dataframe based on 'order_id'
## df_merged_large = df_ords merged with df_ords_prior


df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)
df_merged_large.head()

## Checking full match of merge function
## value_counts() provides a frequency count to see if 100% of 'both's" or if there are any left/rights


df_merged_large['_merge'].value_counts()

## There are 0 left or right_only's so we can conclude a 100% match. HOWEVER, it is important to understand the nature of the joins used. If we used am "outer" join, it will show a left_only value of 206209. This indicates there are some rows of data that did not have a match. Inner join allowed only those that had a match to properly merge

## Exporting Data in Pickle Format

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', '4.6 orders_products_combined.pkl'))
