==> Instacart Basket Analysis/Scripts/5 Combining & Exporting Data Part B.ipynb <==
## 4.6 Exercise Questions

# Import libraries

import pandas as pd
import numpy as np
import os

##Importing path, and csv files of products

path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'

# To import 4.6 orders_products_combined.pkl as df_orders_products_combined

df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.6 orders_products_combined.pkl'))

df_orders_products_combined.head()

## 4. Check the shape of the imported dataframe (it should be the same as the one you exported—always check!).

df_orders_products_combined.shape
##(32434489 rows, 12 columns) and is the same size as imported

## Combining orders_products_combined data set with df_prods_checked

## 5. Determine a suitable way to combine the orders_products_combined dataframe with your products data set. Make sure you’re using your wrangled, cleaned, and deduped products data set stored in your “Prepared Data” folder from the previous Exercise’s task.

##Importing products data set from the prepared data folder

df_prods_checked = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', '4.5 prods_checked.csv'), index_col=False)
df_prods_checked.head()



df_prods_checked['prices'].max()

## These two data frames share the product_id columns as a key so that will be used throughout the merge process

df_orders_prod_merged = df_orders_products_combined.merge(df_prods_checked, on = 'product_id', indicator = 'match')

df_orders_prod_merged.head()

## 6. Confirm the results of the merge using the merge flag.

df_orders_prod_merged['_merge'].value_counts()

#There are 0 values of left/right's only inferring everything has been matched accordingly. If this was a different join like an "outer" there may be a larger amonut of values that did not match fully. The inner join nature allows proper matching throughout the merge

# drop the flag column '_merge' from the previous merge, along with unnamed columns for data to look more "clean"

orders_prod_merged_dropped = df_orders_prod_merged.drop(['match', '_merge', 'Unnamed: 0.1', 'Unnamed: 0_x', 'Unnamed: 0_y'], axis = 1 )

orders_prod_merged_dropped.head()
## New DF "orders_prod_merged_dropped" made to prevent over writting in case of error

## 7. Export the newly created dataframe as ords_prods_merge in a suitable format (taking into consideration the size).

## Will use pickle format due to file size

orders_prod_merged_dropped.to_pickle(os.path.join(path, '02 Data','Prepared Data', '4.6 orders_prod_merged.pkl'))

## Will also use csv in case needed in the future
orders_prod_merged_dropped.to_csv(os.path.join(path, '02 Data','Prepared Data', '4.6 orders_prod_merged.csv'))
