==> Instacart Basket Analysis/Scripts/6 Grouping Data & Aggregating Variables Part A.ipynb <==
## 4.8 Grouping Data & Aggregating Variables Part A

##Importing important packages

import pandas as pd
import numpy as np
import os

##Importing path, and csv files of products

path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'

# To import 4.7 orders_prod_merged.pkl as ords_prods_merge

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.7 ords_prods_merged.pkl'))
ords_prods_merge.head()

# Making a subset of the first 1 million rows and using shape function to confirm
df= ords_prods_merge[:1000000]
df.shape

df.head(10)

#Using group by on the product_name column
df.groupby('product_name')

#Running group by with an aggregate function on department_id
# Or can also use: 
# df.groupby('department_id').agg({'order_number': ['mean']})

df.groupby('department_id')['order_number'].mean()

#Aggregating based on multiple arguments

df.groupby('department_id').agg({'order_number':['mean','min','max']})

## Generating a customer loyalty flag

# Setting group by of the user_id column and applying the transform function over order_number

ords_prods_merge['max_order']= ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head()

#Using Loc()funciton for condition if statements

ords_prods_merge.loc[ords_prods_merge['max_order']>40, 'loyalty_flag']= 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order']<= 40) & (ords_prods_merge['max_order']>10), 'loyalty_flag']= 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order']<=10, 'loyalty_flag']= 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

ords_prods_merge.head()

## Exporting dataframe as a pickle file.

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', '4.8 ords_prods_merged part A.pkl'))
