==> Instacart Basket Analysis/Scripts/7 Grouping Data & Aggregating Variables Part B.ipynb <==
## 4.8 Grouping Data & Aggregating Variables Part B

##Importing important packages

import pandas as pd
import numpy as np
import os

## 1. Create a new notebook for this task. Be sure to import the relevant libraries, along with your ords_prods_merge dataframe, which should include your newly derived columns from the previous Exercise.

##Importing path, and csv files of products

path= r'/Users/erictam/Downloads/CareerFoundry Program/Data Immersion Achievement 4/Instacart Basket Analysis'

# To import 4.8 ords_prods_merged part A.pkl as ords_prods_merge

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.8 ords_prods_merged part A.pkl'))
ords_prods_merge.head()

## 2. In this Exercise, you learned how to find the aggregated mean of the “order_number” column grouped by “department_id” for a subset of your dataframe. Now, repeat this process for the entire dataframe.

ords_prods_merge.groupby('department_id').agg({'order_number':['mean']})

## 3. Analyze the result. How do the results for the entire dataframe differ from those of the subset? Include your comments in a markdown cell below the executed code.

Comparing the subset and entire dataframe:

General Trends: The entire dataframe generally has higher mean order numbers across departments, indicating it captures a broader range of customer behaviours than the subset.

Overall, it is important to know that this is a surface level comparison, that only uses the eye test. A proper t-test analysis should be done to understand whether there is a true difference between the mean values

From previous subset
department_id
1     14.805603
2     17.091743
3     17.933180
4     17.901477
5     15.211405
6     15.382228
7     17.700476
8     16.485269
9     15.967282
10    20.091818
11    16.484907
12    15.620661
13    16.490636
14    17.502079
15    15.695987
16    18.017144
17    16.155822
18    19.606536
19    17.634215
20    17.145616
21    22.014421

## 4. Follow the instructions in the Exercise for creating a loyalty flag for existing customers using the transform() and loc() functions.

# This section was done in Part A, with the exercise instructions. From part A, it was exported and later imported into this achievement exercise. Do avoid running the same code, and exporting the same column, it will not be run on this section

#Below is the code that was run in order to achieve the loyalty flag columns

#Using Loc()funciton for condition if statements
#ords_prods_merge.loc[ords_prods_merge['max_order']>40, 'loyalty_flag']= 'Loyal customer'

#ords_prods_merge.loc[(ords_prods_merge['max_order']<= 40) & (ords_prods_merge['max_order']>10), 'loyalty_flag']= 'Regular customer'

#ords_prods_merge.loc[ords_prods_merge['max_order']<=10, 'loyalty_flag']= 'New customer'

#ords_prods_merge['loyalty_flag'].value_counts(dropna= False)

#ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

## 5. The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

# The three types of customers are Loyal customer, Regular customer, New customer
# Group by the loyalty flag, and aggragate the data based on mean prices. This will identify the prices of products purchased by the varying tiers of customers

ords_prods_merge.head()


#Grouping by and aggragating by loyalty flag
ords_prods_merge.groupby('loyalty_flag').agg({'order_number':['mean','median','min','max']})

## 6. The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria:

# a. If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.”

# b. If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# Will need to group by the user_id and transpose the average of product prices. 
# Then will need to use loc() function to extablish the spending flag based on the given criteria

ords_prods_merge['Avg_Spent']= ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)
ords_prods_merge.head()

#Using loc() function to establish the spending flag

ords_prods_merge.loc[ords_prods_merge['Avg_Spent']<10,'spender_type']='Low Spender'
ords_prods_merge.loc[ords_prods_merge['Avg_Spent']>=10,'spender_type']='High Spender'
ords_prods_merge['spender_type'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'Avg_Spent','spender_type']].head()

## 7. In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:

# a. If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
# b. If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
# c. If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# Will need to group by the user_id and transpose the median of days_since_prior_order data. 
# Then will need to use loc() function to extablish the frequency flag based on the given criteria

ords_prods_merge['median_order_frequency']= ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)
ords_prods_merge.head()

# using loc() function to establish frequency flag

ords_prods_merge.loc[ords_prods_merge['median_order_frequency'] >20, 'frequency_flag']= 'Non-Frequent customer'
ords_prods_merge.loc[(ords_prods_merge['median_order_frequency']<=20) & (ords_prods_merge['median_order_frequency']>10), 'frequency_flag']= 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['median_order_frequency']<=10, 'frequency_flag']= 'Frequent customer'

ords_prods_merge['frequency_flag'].value_counts(dropna= False)

ords_prods_merge[['user_id', 'median_order_frequency','frequency_flag']].head()

## Exporting dataframe as a pickle file.

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', '4.8 ords_prods_merged part B.pkl'))
